+++ app/brand-assets/opengraph-image.alt.txt
@@ -1,0 +1,1 @@
+NameHash Labs - Brand Assets
