+++ app/contact/opengraph-image.alt.txt
@@ -1,0 +1,1 @@
+NameHash Labs - Contact Us
