+++ app/ens-referral-program/opengraph-image.alt.txt
@@ -1,0 +1,1 @@
+ENS Referral Program Proposal
