+++ app/layout.tsx
@@ -1,0 +1,72 @@
+import "./globals.css";
+import "@namehash/ens-webfont";
+
+import type { Metadata } from "next";
+import { Metadata as NamehashMetadata } from "@namehash/utils";
+import { Inter } from "next/font/google";
+import { SpeedInsights } from "@vercel/speed-insights/next";
+import { Analytics } from "@vercel/analytics/react";
+
+import { HeadlineBanner } from "@/components/1 - atoms";
+import { Header } from "@/components/2 - molecules/header";
+import { Footer } from "@/components/2 - molecules";
+
+const inter = Inter({
+  variable: "--font-inter",
+  subsets: ["latin"],
+});
+
+const title = "NameHash Labs - Helping ENS Grow for ENS";
+const description =
+  "NameHash Labs builds open source public goods that drive the global adoption of ENS.";
+const keywords = ["ens", "web3", "eth", "nameguard", "namekit", "namehash"];
+
+export const metadata: Metadata = {
+  ...NamehashMetadata.defaultMetdata,
+  metadataBase: new URL("https://namehashlabs.org"),
+  alternates: {
+    canonical: "/",
+  },
+  title: {
+    template: "NameHash Labs - %s",
+    default: title,
+  },
+  description,
+  keywords,
+  openGraph: {
+    ...NamehashMetadata.defaultMetdata.openGraph,
+    title,
+    description,
+  },
+  twitter: {
+    ...NamehashMetadata.defaultMetdata.twitter,
+    description,
+  },
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body className={`${inter.variable} font-sans`}>
+        <div className="w-full overflow-hidden">
+          <HeadlineBanner />
+          <div className="w-full flex flex-col relative items-center justify-center">
+            <div className="absolute top-0 left-0 w-full z-40">
+              <Header />
+            </div>
+
+            {children}
+          </div>
+
+          <Footer />
+        </div>
+        <SpeedInsights />
+        <Analytics />
+      </body>
+    </html>
+  );
+}
