+++ app/page.tsx
@@ -1,0 +1,37 @@
+import { JoinOurTeamSection } from "@/components/1 - atoms/join-our-team-section";
+import {
+  HeroSection,
+  OurPrinciplesSection,
+  ProductsSection,
+  TeamSection,
+} from "@/components/2 - molecules";
+import { FinancialSupportSection } from "@/components/2 - molecules/financial-support-section";
+import { IntroducingNamekit } from "@/components/2 - molecules/introducing-namekit";
+import { OurSuportersSection } from "@/components/2 - molecules/our-supporters-section";
+import { TestimonialsSection } from "@/components/2 - molecules/testimonials-section";
+import { TheVisionSection } from "@/components/2 - molecules/the-vision-section";
+import { type Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "NameHash Labs - Helping ENS Grow",
+  description:
+    "NameHash Labs builds open source public goods that drive the global adoption of ENS.",
+  keywords: ["nameguard", "normalization", "ens", "web3", "eth"],
+};
+
+export default function Page() {
+  return (
+    <main className="w-full flex flex-col items-center justify-between">
+      <HeroSection />
+      <TheVisionSection />
+      <IntroducingNamekit />
+      <ProductsSection />
+      <OurSuportersSection />
+      <TestimonialsSection />
+      <OurPrinciplesSection />
+      <TeamSection />
+      <JoinOurTeamSection />
+      <FinancialSupportSection />
+    </main>
+  );
+}
