+++ app/partners/twitter-image.png
