+++ package.json
@@ -7,34 +7,35 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@adraffy/ens-normalize": "^1.10.1",
-    "@calcom/embed-react": "^1.3.0",
-    "@heroicons/react": "^2.0.18",
+    "@adraffy/ens-normalize": "1.10.1",
+    "@calcom/embed-react": "1.4.0",
+    "@heroicons/react": "2.1.3",
     "@namehash/ens-webfont": "0.0.3",
-    "@vercel/analytics": "^1.1.1",
-    "@vercel/speed-insights": "^1.0.9",
-    "classcat": "^5.0.4",
-    "fast-average-color": "^9.4.0",
-    "lottie-react": "^2.4.0",
-    "next": "14.0.1",
+    "@namehash/utils": "^0.0.1",
+    "@vercel/analytics": "1.2.2",
+    "@vercel/speed-insights": "1.0.10",
+    "classcat": "5.0.5",
+    "fast-average-color": "9.4.0",
+    "lottie-react": "2.4.0",
+    "next": "14.2.2",
     "next-seo": "^6.5.0",
-    "nextjs-cors": "^2.2.0",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-tooltip": "^5.25.1",
-    "react-wrap-balancer": "^1.1.0",
-    "swiper": "^11.0.5",
-    "yup": "^1.3.3"
+    "nextjs-cors": "2.2.0",
+    "react": "18",
+    "react-dom": "18",
+    "react-tooltip": "5.26.3",
+    "react-wrap-balancer": "1.1.0",
+    "swiper": "11.1.1",
+    "yup": "1.4.0"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.0.1",
-    "eslint": "^8",
-    "eslint-config-next": "14.0.1",
-    "postcss": "^8",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "@types/node": "20.12.7",
+    "@types/react": "18.2.79",
+    "@types/react-dom": "18.2.25",
+    "autoprefixer": "10.4.19",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^14.2.2",
+    "postcss": "8.4.38",
+    "tailwindcss": "3.4.3",
+    "typescript": "5.4.5"
   }
 }
